requirements.txt
alembic==1.15.2
annotated-types==0.7.0
anyio==4.9.0
blinker==1.9.0
cffi==1.17.1
chardet==5.2.0
click==8.1.8
colorama==0.4.6
cryptography==44.0.2
distlib==0.3.9
dotenv==0.9.9
fastapi==0.115.12
filelock==3.17.0
Flask==3.1.0
Flask-Login==0.6.3
Flask-Migrate==4.1.0
Flask-Script==2.0.6
Flask-SQLAlchemy==3.1.1
greenlet==3.1.1
gunicorn==23.0.0
h11==0.16.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
Mako==1.3.9
MarkupSafe==3.0.2
mysql-connector==2.2.9
mysql-connector-python==9.2.0
mysql-connector-python-rf==2.2.2
mysqlclient==2.2.7
packaging==24.2
pillow==11.1.0
platformdirs==4.3.6
psycopg2==2.9.10
pycparser==2.22
pydantic==2.11.4
pydantic_core==2.33.2
pygame==2.6.1
PyMySQL==1.1.1
python-dotenv==1.1.0
python-multipart==0.0.20
reportlab==4.3.1
sniffio==1.3.1
SQLAlchemy==2.0.39
starlette==0.46.2
tailwind==3.1.5b0
typing-inspection==0.4.0
typing_extensions==4.13.0
uvicorn==0.34.2
virtualenv==20.29.2
waitress==3.0.2
Werkzeug==3.1.3
